refactor: Organize util functions into dedicated files, clean up Signature Insight
