+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/hashdit/metamask-snap.git"
   },
   "source": {
-    "shasum": "yvagkTYHIyuiyCmlBbqruHkkYzuR7rko2qYSq1+oooY=",
+    "shasum": "DmD41++k1YCRx6wDWP5RZSyh74kXQ1a/A8XHhsOqd14=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
