+++ packages/snap/src/utils/utils.ts
@@ -3,18 +3,7 @@ import { v4 as uuidv4 } from 'uuid';
 import hmacSHA256 from 'crypto-js/hmac-sha256';
 import encHex from 'crypto-js/enc-hex';
 import { CHAINS_INFO } from './chains';
-import {
-	heading,
-	panel,
-	text,
-	copyable,
-	divider,
-	address,
-	row,
-	Signature,
-} from '@metamask/snaps-sdk';
-import { errorContent } from './content';
-import { keccak256 } from 'js-sha3';
+
 
 // Called during HashDit Snap installation. Used to authenticate the user with DiTing, and retrieve an API key.
 export async function authenticateDiTing(
@@ -301,104 +290,6 @@ export async function isEOA(address: any) {
 	}
 }
 
-// Perform similarity score to detect address poisoning attacks
-export function addressPoisoningDetection(
-	userAddresses: string[],
-	targetAddresses: string[],
-) {
-	let resultArray: any[] = [];
-	let similarityResult = detectSimilarity(userAddresses, targetAddresses);
-	if (similarityResult.length > 0) {
-		resultArray.push(
-			heading('Address Poisoning'),
-			text(
-				`You are about to interact with an address that appears similar to one of your personal addresses. This could be an attempt to steal your funds. Please verify the addresses before proceeding.`,
-			),
-		);
-		for (var i = 0; i < similarityResult.length; i++) {
-			resultArray.push(
-				row('Your Address', address(similarityResult[i].userAddress)),
-				row(
-					'Similar Address',
-					address(similarityResult[i].targetAddress),
-				),
-				row(
-					'Risk Level',
-					text(`${similarityResult[i].similarityRiskLevel}`),
-				),
-				divider(),
-			);
-		}
-	}
-	return resultArray;
-}
-
-/**
- * The function compares the first and last 5 hexadecimals of two Ethereum addresses.
- * It assesses their prefix and postfix similarity and returns a score ranging from 0 (no similarity) to 5 (complete match).
- * Score increments only when both the 1st character of prefix & suffix match the target address.
- * Skip if the addresses are the same.
- */
-function detectSimilarity(
-	userAddressArray: string[],
-	targetAddressArray: string[],
-) {
-	var similarityScoreResultArray = [];
-
-	for (let userAddress of userAddressArray) {
-		for (let targetAddress of targetAddressArray) {
-			// Only compare the addresses after the `0x` prefix
-			// Set to lowercase for consistency
-			const userAddressConvert = userAddress.toLowerCase().substring(2);
-			const targetAddressCovert = targetAddress
-				.toLowerCase()
-				.substring(2);
-
-			// Addresses are identical. Don't need to consider similarity.
-			if (userAddressConvert == targetAddressCovert) {
-				continue;
-			}
-
-			let similarityScore = 0;
-			const addressLength = 39;
-
-			// Compare first 5 hex
-			for (var i = 0; i < 5; i++) {
-				if (
-					userAddressConvert[i] == targetAddressCovert[i] &&
-					userAddressConvert[addressLength - i] ==
-						targetAddressCovert[addressLength - i]
-				) {
-					similarityScore += 1;
-				}
-			}
-
-			// If there are more than 3 matching prefix or postfix characters, we send a warning to the user.
-			if (similarityScore >= 3) {
-				let similarityRiskLevel;
-				switch (similarityScore) {
-					case 3:
-						similarityRiskLevel = '⛔ High Risk ⛔';
-						break;
-					case 4:
-						similarityRiskLevel = '⛔ High Risk ⛔';
-						break;
-					case 5:
-						similarityRiskLevel = '🚫 **Critical Risk** 🚫';
-						break;
-				}
-
-				similarityScoreResultArray.push({
-					userAddress,
-					targetAddress,
-					similarityRiskLevel,
-				});
-			}
-		}
-	}
-	return similarityScoreResultArray;
-}
-
 // Determine the risk title and description for each risk level. Used by URL screening.
 function determineUrlRiskInfo(urlRiskLevel: number): string[] {
 	if (urlRiskLevel == 0) {
@@ -455,290 +346,6 @@ export function determineTransactionAndDestinationRiskInfo(riskLevel: number) {
 		return [
 			'Low',
 			'This transaction is considered low risk. Please review the details of this transaction.',
-		];
-	} else {
-		return [
-			'Unknown',
-			'The risk level of this transaction is unknown. Please proceed with caution.',
-		];
-	}
-}
-
-// Expected structure for EIP712 signature
-export interface SignatureParsed {
-	from: string;
-	data: {
-		types: {
-			EIP712Domain: Array<{ name: string; type: string }>;
-			PermitBatch?: Array<{ name: string; type: string }>;
-			PermitDetails?: Array<{ name: string; type: string }>;
-			PermitSingle?: Array<{ name: string; type: string }>;
-			PermitForAll?: Array<{ name: string; type: string }>;
-			Permit?: Array<{ name: string; type: string }>;
-		};
-		domain: {
-			name: string;
-			chainId: string;
-			verifyingContract: string;
-		};
-		primaryType: string;
-		message: any;
-	};
-}
-
-// Determine if the signature is a Permit signature.
-export async function parseSignature(
-	signature: Signature,
-	signatureOrigin: any,
-) {
-	let signatureParsed: SignatureParsed;
-	let decodedData;
-
-	// Check if signature.data is an object. Exit if data is not an object
-	if (typeof signature.data !== 'object') {
-		//console.log('Invalid data type for signature.data:', typeof signature.data);
-		return null;
-	}
-
-	decodedData = signature.data;
-
-	signatureParsed = {
-		from: signature.from,
-		data: decodedData,
-	};
-
-	const { primaryType, message } = signatureParsed.data;
-	//console.log('PrimaryType and Message', primaryType, message);
-
-	let spender: string | undefined;
-	let token: any;
-	let amount: string | undefined;
-
-	// Check for PermitBatch
-	if (
-		primaryType === 'PermitBatch' &&
-		message.details &&
-		Array.isArray(message.details)
-	) {
-		spender = message.spender;
-		token = message.details;
-		amount = undefined;
-	}
-	// Check for PermitSingle
-	else if (primaryType === 'PermitSingle' && message.details) {
-		spender = message.spender;
-		token = message.details.token;
-		amount = message.details.amount;
-	}
-	// Check for PermitForAll
-	else if (primaryType === 'PermitForAll') {
-		spender = message.operator;
-		token = undefined; // Not relevant for PermitForAll
-		amount = undefined; // Not relevant for PermitForAll
-	}
-	// Check for Permit (ERC20/other token)
-	else if (primaryType === 'Permit' && message.spender) {
-		spender = message.spender;
-		amount = message.value;
-
-		if (message.token) {
-			token = message.token;
-		} else {
-			token = null;
-		}
-	} else {
-		//console.log('Not a Permit Signature, returning null');
-
-		return null;
-	}
-
-	return await callHashDitAPIForSignatureInsight(
-		primaryType,
-		spender,
-		token,
-		signatureOrigin,
-	);
-}
-
-// Call the HashDit API to retrieve risk levels
-async function callHashDitAPIForSignatureInsight(
-	primaryType: any,
-	spender: any,
-	tokenAddress: any,
-	signatureOrigin: any,
-) {
-	let contentArray: any[] = [];
-	let persistedUserData;
-	// Retrieve persisted user data
-	try {
-		persistedUserData = await snap.request({
-			method: 'snap_manageState',
-			params: { operation: 'get' },
-		});
-	} catch (error) {
-		console.error('Error retrieving persisted user data:', error);
-		contentArray = errorContent;
-		return contentArray;
-	}
-
-	if (persistedUserData !== null) {
-		// Set up the URL detection request promise. This is executed for all permit types.
-		const urlRequestPromise = getHashDitResponse(
-			'hashdit_snap_tx_api_url_detection',
-			persistedUserData,
-			signatureOrigin,
-		);
-
-		// Set up additional promises depending on the primaryType
-		const blacklistPromises: Array<Promise<any>> = [];
-		const chainId = await ethereum.request({ method: 'eth_chainId' });
-		if (
-			typeof chainId === 'string' &&
-			(chainId === '0x38' || chainId === '0x1')
-		) {
-			if (primaryType === 'Permit' || primaryType === 'PermitSingle') {
-				// Call blacklist API on spender and token address
-				blacklistPromises.push(
-					getHashDitResponse(
-						'signature_insight_blacklist',
-						persistedUserData,
-						null,
-						null,
-						chainId,
-						spender,
-					),
-				);
-			} else if (primaryType === 'PermitForAll') {
-				blacklistPromises.push(
-					getHashDitResponse(
-						'signature_insight_blacklist',
-						persistedUserData,
-						null,
-						null,
-						chainId,
-						spender,
-					),
-				);
-			} else if (primaryType === 'PermitBatch') {
-				blacklistPromises.push(
-					getHashDitResponse(
-						'signature_insight_blacklist',
-						persistedUserData,
-						null,
-						null,
-						chainId,
-						spender,
-					),
-				);
-			}
-		}
-
-		// Resolve all promises, including the URL request and blacklist requests concurrently
-		const responses = await Promise.all([
-			urlRequestPromise,
-			...blacklistPromises,
-		]);
-		return await createContentForSignatureInsight(
-			responses,
-			spender,
-			signatureOrigin,
-		);
-	}
-}
-
-// Take the API results and create the contentArrays returned to onSignature() function.
-async function createContentForSignatureInsight(
-	responses: any,
-	spender: any,
-	signatureOrigin: string,
-) {
-	// Define variables for responses
-	let urlResp, spenderBlacklistResp, tokenBlacklistResp, isSpenderEOA;
-	let contentArray: any[] = [];
-
-	// Assign responses based on primaryType
-	urlResp = responses[0];
-	spenderBlacklistResp = responses[1];
-
-	// Check if spender is an Externally Owned Address.
-	// We consider approving to an EOA spender to be a high risk because there are few scenarios where this is needed.
-	isSpenderEOA = await isEOA(spender);
-	// If spender is EOA and spender isn't blacklisted
-	if (isSpenderEOA) {
-		// If the chain is not ETH or BSC, then spenderBlacklistResp will be undefined
-		if (
-			spenderBlacklistResp == undefined ||
-			spenderBlacklistResp.overall_risk <= 2
-		) {
-			contentArray.push(
-				heading('Spender Screening'),
-				row('Risk Level', text('⛔ High Risk ⛔')),
-				row('Spender', address(spender)),
-				text(
-					'This transaction’s spender is an Externally Owned Account (EOA), likely indicating a scam. Approving it will give a third-party direct access to your funds, risking potential loss. It is advised to reject this transaction.',
-				),
-				divider(),
-				heading('Website Screening'),
-				row('Website', text(signatureOrigin)),
-				row('Risk Level', text(urlResp.url_risk_level)),
-				text(urlResp.url_risk_detail),
-			);
-			return contentArray;
-		}
-	}
-
-	// Supported chain
-	if (spenderBlacklistResp != undefined) {
-		if (spenderBlacklistResp.overall_risk != -1) {
-			// Convert risk level to risk title
-			const [riskTitle, riskOverview] = determineSpenderRiskInfo(
-				spenderBlacklistResp.overall_risk,
-			);
-			contentArray.push(
-				heading('Spender Screening'),
-				row('Risk Level', text(riskTitle)),
-				row('Spender', address(spender)),
-				text(riskOverview),
-				divider(),
-				heading('Website Screening'),
-				row('Website', text(signatureOrigin)),
-				row('Risk Level', text(urlResp.url_risk_level)),
-				text(urlResp.url_risk_detail),
-			);
-
-			return contentArray;
-		}
-	}
-	// Fallback content value if (not a supported chain and spender is not EOA), or (API return overall_risk level of -1 / unknown)
-
-	contentArray.push(
-		heading('Spender Screening'),
-		row('Risk Level', text('Low')),
-		row('Spender', address(spender)),
-		text(
-			'This transaction’s spender is not blacklisted by HashDit. However, approving it will give a third-party direct access to your funds, risking potential loss. Please proceed with caution. Default risk level.',
-		),
-		divider(),
-		heading('Website Screening'),
-		row('Website', text(signatureOrigin)),
-		row('Risk Level', text(urlResp.url_risk_level)),
-		text(urlResp.url_risk_detail),
-	);
-
-	return contentArray;
-}
-
-// Determine the risk title and description for each risk level. Used by Signature Insight.
-export function determineSpenderRiskInfo(riskLevel: number) {
-	if (riskLevel >= 2) {
-		return [
-			'⛔ High ⛔',
-			'This transaction’s spender is **blacklisted** by HashDit, likely indicating a scam. Approving it will give a third-party direct access to your funds, risking potential loss. It is advised to reject this transaction.',
-		];
-	} else if (riskLevel >= 0) {
-		return [
-			'Low',
-			'This transaction’s spender is not blacklisted by HashDit. However, approving it will give a third-party direct access to your funds, risking potential loss. Please proceed with caution. Default risk level.',
 		];
 	} else {
 		return [
